Assigning loot ownership + small fixes (#181)
* Added loot ownership

* Map items can be spawned with an owner, and only be picked up by the owner and their party. Anyone can still see it, however.
* The time that loot is exclusive to the owner and their party is configurable through the server's config.json under Map Options.
* NPC Drops are owned by the player (and their party) that dealt the most damage to them. (Unless killed by an NPC or something else, then there is no owner)
* Player drops on death are owned by whoever dealt the killing blow and their party.
* Items dropped from your inventory to the map are owned by you and your party until the timer expires.

* Made the packet handler for PickupItempacket slightly less a hassle to read by shortening some checks.

* Removed useless reference and fixed an issue

Removed a useless reference to System from a random test I forgot to remove.

Fixed an oversight in the way the owner was determined by the amount of damage done to an NPC.. It should now no longer always use the last player in the list, but the one with the most damage.

* General code cleanup

Cleaned up my loot tagging attempt in Entity.Die, as well as some general cleaning up of repeated code using Npc.DamageBase.

* Fixed an NPC behaviour check I skipped over last cleanup

* Add an option to hide loot for players until the ownership claim drops.

Added a configurable option to the server configuration (ShowUnownedItems) that can either show or hide loot for players that are not the owner and not in the owner's party until the claim on that item is dropped by the server.

Disabled attempts at picking up invisible items as well.

* Fix a potential error here, in case the item gets deleted before it should become visible.

* Small death rule change if player dies from anything but another player.

Changed the way item ownership is handled upon player death to set ownership to them (and their party) should a player die from anything that is not another player.

* Moved the two new options to their own Options section.

Would prefer to move these as well:

ItemDespawnTime
ItemSpawnTime

But this would definitely break existing configs.

* Added the error a player gets when attempting to pick up an item that does not belong to them localizable.

* Requested changes for pr #181

- Added Xml summaries to LootOptions.cs
- Made LootOptions a child of Options rather than adding more top level options.
- Changed both Player.IsInMyParty checks to instead be simple Linq queries and enforces a player to be passed on.
- Fixed some places that would throw any entity into the Player.IsInMyParty method.
- Changed the new translation string to be Mixed Case and added the NotNull property.
- Changed some MapItem update logic on MapInstance to not refer to each item by id every time, but a simpler re-usable variable.
- Changed more Guid comparisons
- Added a message to picking up items on the server side if the user's inventory is full.
- Added some Todo's to MapOptions for items that could be moved or renamed for clarity.

* Updated the alert messages to utilize CustomColors

* JC Said I could do this. :coy:

Removes ItemDespawnTime from the MapOptions class and places it in the LootOptions class where it beloongs.

Renames ItemSpawnTime in MapOptions to ItemAttributeRespawnTime which is far more descriptive of what it actually does.

Removes these two options from the top level Options class, like why is almost everything there anyway if we have categories?

* Items forcibly returned by trading and items dropped by resources are now tagged with an owner.

* Fix a small bug unrelated to my code related to regenerating NPC vitals, the code would never actually execute for NPCs.

* Resolve some review problems.

Signed-off-by: Greg Plachno <gplachno@ycp.edu>